Separate network specific data serialiser traits
