+++ src/openrct2/core/DataSerialiserTraits.h
@@ -14,13 +14,11 @@
 #include "../core/MemoryStream.h"
 #include "../localisation/Localisation.h"
 #include "../network/NetworkTypes.h"
-#include "../network/network.h"
 #include "../object/Object.h"
 #include "../ride/Ride.h"
 #include "../ride/TrackDesign.h"
 #include "../world/Location.hpp"
 #include "../world/TileElement.h"
-#include "DataSerialiserTag.h"
 #include "Endianness.h"
 #include "MemoryStream.h"
 
@@ -178,125 +176,6 @@ template<> struct DataSerializerTraits_t<std::string>
     }
 };
 
-template<> struct DataSerializerTraits_t<NetworkPlayerId_t>
-{
-    static void encode(OpenRCT2::IStream* stream, const NetworkPlayerId_t& val)
-    {
-        uint32_t temp = static_cast<uint32_t>(val.id);
-        temp = ByteSwapBE(temp);
-        stream->Write(&temp);
-    }
-    static void decode(OpenRCT2::IStream* stream, NetworkPlayerId_t& val)
-    {
-        uint32_t temp;
-        stream->Read(&temp);
-        val.id = static_cast<decltype(val.id)>(ByteSwapBE(temp));
-    }
-    static void log(OpenRCT2::IStream* stream, const NetworkPlayerId_t& val)
-    {
-        char playerId[28] = {};
-        snprintf(playerId, sizeof(playerId), "%u", val.id);
-
-        stream->Write(playerId, strlen(playerId));
-
-        int32_t playerIndex = OpenRCT2::GetContext()->GetNetwork()->GetPlayerIndex(val.id);
-        if (playerIndex != -1)
-        {
-            const char* playerName = OpenRCT2::GetContext()->GetNetwork()->GetPlayerName(playerIndex);
-            if (playerName != nullptr)
-            {
-                stream->Write(" \"", 2);
-                stream->Write(playerName, strlen(playerName));
-                stream->Write("\"", 1);
-            }
-        }
-    }
-};
-
-template<> struct DataSerializerTraits_t<NetworkRideId_t>
-{
-    static void encode(OpenRCT2::IStream* stream, const NetworkRideId_t& val)
-    {
-        uint32_t temp = static_cast<uint32_t>(val.id);
-        temp = ByteSwapBE(temp);
-        stream->Write(&temp);
-    }
-    static void decode(OpenRCT2::IStream* stream, NetworkRideId_t& val)
-    {
-        uint32_t temp;
-        stream->Read(&temp);
-        val.id = static_cast<decltype(val.id)>(ByteSwapBE(temp));
-    }
-    static void log(OpenRCT2::IStream* stream, const NetworkRideId_t& val)
-    {
-        char rideId[28] = {};
-        snprintf(rideId, sizeof(rideId), "%u", val.id);
-
-        stream->Write(rideId, strlen(rideId));
-
-        auto ride = get_ride(val.id);
-        if (ride != nullptr)
-        {
-            auto rideName = ride->GetName();
-
-            stream->Write(" \"", 2);
-            stream->Write(rideName.c_str(), rideName.size());
-            stream->Write("\"", 1);
-        }
-    }
-};
-
-template<typename T> struct DataSerializerTraits_t<DataSerialiserTag<T>>
-{
-    static void encode(OpenRCT2::IStream* stream, const DataSerialiserTag<T>& tag)
-    {
-        DataSerializerTraits<T> s;
-        s.encode(stream, tag.Data());
-    }
-    static void decode(OpenRCT2::IStream* stream, DataSerialiserTag<T>& tag)
-    {
-        DataSerializerTraits<T> s;
-        s.decode(stream, tag.Data());
-    }
-    static void log(OpenRCT2::IStream* stream, const DataSerialiserTag<T>& tag)
-    {
-        const char* name = tag.Name();
-        stream->Write(name, strlen(name));
-        stream->Write(" = ", 3);
-
-        DataSerializerTraits<T> s;
-        s.log(stream, tag.Data());
-
-        stream->Write("; ", 2);
-    }
-};
-
-template<> struct DataSerializerTraits_t<OpenRCT2::MemoryStream>
-{
-    static void encode(OpenRCT2::IStream* stream, const OpenRCT2::MemoryStream& val)
-    {
-        DataSerializerTraits<uint32_t> s;
-        s.encode(stream, val.GetLength());
-
-        stream->Write(val.GetData(), val.GetLength());
-    }
-    static void decode(OpenRCT2::IStream* stream, OpenRCT2::MemoryStream& val)
-    {
-        DataSerializerTraits<uint32_t> s;
-
-        uint32_t length = 0;
-        s.decode(stream, length);
-
-        std::unique_ptr<uint8_t[]> buf(new uint8_t[length]);
-        stream->Read(buf.get(), length);
-
-        val.Write(buf.get(), length);
-    }
-    static void log(OpenRCT2::IStream* stream, const OpenRCT2::MemoryStream& tag)
-    {
-    }
-};
-
 template<typename _Ty, size_t _Size> struct DataSerializerTraitsPODArray
 {
     static void encode(OpenRCT2::IStream* stream, const _Ty (&val)[_Size])
@@ -442,6 +321,32 @@ template<typename _Ty> struct DataSerializerTraits_t<std::vector<_Ty>>
     }
 };
 
+template<> struct DataSerializerTraits_t<OpenRCT2::MemoryStream>
+{
+    static void encode(OpenRCT2::IStream* stream, const OpenRCT2::MemoryStream& val)
+    {
+        DataSerializerTraits<uint32_t> s;
+        s.encode(stream, val.GetLength());
+
+        stream->Write(val.GetData(), val.GetLength());
+    }
+    static void decode(OpenRCT2::IStream* stream, OpenRCT2::MemoryStream& val)
+    {
+        DataSerializerTraits<uint32_t> s;
+
+        uint32_t length = 0;
+        s.decode(stream, length);
+
+        std::unique_ptr<uint8_t[]> buf(new uint8_t[length]);
+        stream->Read(buf.get(), length);
+
+        val.Write(buf.get(), length);
+    }
+    static void log(OpenRCT2::IStream* stream, const OpenRCT2::MemoryStream& tag)
+    {
+    }
+};
+
 template<> struct DataSerializerTraits_t<MapRange>
 {
     static void encode(OpenRCT2::IStream* stream, const MapRange& v)
@@ -583,26 +488,6 @@ template<> struct DataSerializerTraits_t<CoordsXYZD>
         snprintf(
             msg, sizeof(msg), "CoordsXYZD(x = %d, y = %d, z = %d, direction = %d)", coord.x, coord.y, coord.z, coord.direction);
         stream->Write(msg, strlen(msg));
-    }
-};
-
-template<> struct DataSerializerTraits_t<NetworkCheatType_t>
-{
-    static void encode(OpenRCT2::IStream* stream, const NetworkCheatType_t& val)
-    {
-        uint32_t temp = ByteSwapBE(val.id);
-        stream->Write(&temp);
-    }
-    static void decode(OpenRCT2::IStream* stream, NetworkCheatType_t& val)
-    {
-        uint32_t temp;
-        stream->Read(&temp);
-        val.id = ByteSwapBE(temp);
-    }
-    static void log(OpenRCT2::IStream* stream, const NetworkCheatType_t& val)
-    {
-        const char* cheatName = CheatsGetName(static_cast<CheatType>(val.id));
-        stream->Write(cheatName, strlen(cheatName));
     }
 };
